+++ RELEASE_NOTES.md
@@ -2,6 +2,22 @@
 The Narrative Interface allows users to craft KBase Narratives using a combination of GUI-based commands, Python and R scripts, and graphical output elements.
 
 This is built on the IPython Notebook (more notes will follow).
+### Version 0.3.15 - 1/21/2015
+__Changes__
+- Updated type name for Assembly File transform method
+- Added reset button to App cells (method cells still need fixing)
+- Added widgets for metagenome sets
+__Bugfixes__
+- JIRA NAR-418 - fixed issue where job error was cleared on panel refresh.
+- Fixed issue where method panel error state wasn't being activated properly.
+
+### Version 0.3.14 - 1/21/2015
+__Changes__
+- Updated server-side Narrative management code to always keep a queue of unattached Narrative containers present. When a user logs on, they already have one ready to connect to.
+- Added visualization widget for microbial community abundance and boxplots.
+- Method details and documentation are visible in a new window now.
+- Added app and method icons for the method panel.
+- Added minimize to app & method panels
 
 ### Version 0.3.13 - 1/20/2015
 __Changes__